Fix dynamic thread count (float -> int)

+++ xrLC/Build.cpp
@@ -22,7 +22,7 @@ float					f_lmap_quality = 4.f;
 // KD end
 BOOL					gl_linear	= FALSE;
 
-float					i_ThreadCount = 1;
+int						i_ThreadCount = 1;
 
 //////////////////////////////////////////////////////////////////////
