+++ xrLC/xrDO_Light/global_options.h
@@ -2,4 +2,4 @@
 
 extern BOOL						b_norgb;
 extern BOOL						b_nosun;
-extern float					i_ThreadCount;
+extern int						i_ThreadCount;
